Add support for dotCover attribute filters.
ExcludeFromCodeCoverage is only available in .NET. I'd like to use a
similar attribute in some .NET 3.5 code, and this change allows me to do
that.

I've also renamed the previous `DotNetFilters` parameter to `Filters`,
alongside the new `AttributeFilters` parameter. The previous name was
tautological, and the new names more closely reflect the dotCover
terminology.